Nasty, dirty commit... FIXME
Need a lot of investigation and cleanup. c_s in phase 2 "works".
However, this is only because of a nasty hack in XBar and it has no
business working. It appears that (now that I've experienced it first
hand) unsurprisingly the cell(s) that get passed through the chain of
eval()'s is the pseudo domain cells, and it doesn't get translated
through XBar.